requirements.txt
autogen-agentchat==0.2.40
gradio==5.9.1 
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
numpy==1.26.4
openai==1.58.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.7.0
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.34.0